pyproject.toml
[tool.poetry]
name = "ctf-platform"
package-mode = false
version = "0.1.0"
description = ""
authors = ["Ivan Krasnopolsky <i1.krasnopolskiy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.34.149"
coolname = "^2.2.0"
django = "^5.0.7"
django-admin-searchable-dropdown = "^1.2"
django-cors-headers = "^4.4.0"
django-environ = "^0.11.2"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
factory-boy = "^3.3.0"
gunicorn = "^22.0.0"
mysqlclient = "^2.2.4"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
line_length = 120
profile = "black"
